requirements.txt
python-dotenv>=1.0.0
anthropic>=0.18.1
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.15
langsmith>=0.0.87
tavily-python
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
elevenlabs